+++ module-system/sov-cli/src/workflows/mod.rs
@@ -1,0 +1,3 @@
+//! Workflows for the CLI wallet
+pub mod keys;
+pub mod transactions;
